chore: move tooltips to common folder
