+++ src/core_modules/capture-core/components/Tooltips/ConditionalTooltip/ConditionalTooltip.component.js
