+++ src/core_modules/capture-core/components/WidgetEventEdit/DataEntry/withMainButton.js
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { connect } from 'react-redux';
 import i18n from '@dhis2/d2-i18n';
 import { Button } from '@dhis2/ui';
-import { ConditionalTooltip } from 'capture-core/components/ConditionalTooltip';
+import { ConditionalTooltip } from 'capture-core/components/Tooltips/ConditionalTooltip';
 import { type RenderFoundation } from '../../../metaData';
 
 type Props = {
