+++ src/core_modules/capture-core/components/WorkingLists/WorkingListsBase/TemplateSelectorChipContent.component.js
@@ -2,6 +2,7 @@
 
 type Props = {
     currentListIsModified: boolean,
+    maxCharacters: number,
     text: string,
     isSelectedTemplate: boolean,
 };
@@ -10,10 +11,11 @@ export const TemplateSelectorChipContent = (props: Props) => {
     const {
         currentListIsModified,
         text,
+        maxCharacters,
         isSelectedTemplate,
     } = props;
 
-    const truncatedText = text.length > 30 ? `${text.substring(0, 27)}...` : text;
+    const truncatedText = text.length > maxCharacters ? `${text.substring(0, maxCharacters - 3)}...` : text;
     const content = isSelectedTemplate && currentListIsModified ? `${truncatedText} *` : truncatedText;
 
     return content;
